Analyze the BR affiliation proportion in scl, w/ R
That's a full analysis written as an R Markdown notebook, it's not
ready for exporting to PDF with knitr but exportin to HTML is ok
(though the file will be huge)

The analysis includes an hypothesis check: is the proportion of BR
affiliations decreasing in the research articles from the SciELO
Brazil collection? That means, are these selected journals
receiving/accepting more research articles from foreign
universities/institutes/schools/faculties?